Optimized by caching the inverse matrix of the transform

+++ include/utils/FloatUtils.h
@@ -4,5 +4,5 @@
 
 
 static bool compareFloats(float a, float b) {
-	return fabs(a - b) <= 0.00001f;
+	return fabsf(a - b) <= 0.00001f;
 }
